[ECP-9482] Updating tests
